requirements.txt
# Core LangChain & Google Vertex AI
langchain==0.2.14
langchain-core==0.2.35
langchain-community==0.2.12
langchain-google-vertexai==1.0.5

# Google Cloud & Vertex AI SDKs
google-cloud-aiplatform>=1.71.1
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1

# Vector Stores
faiss-cpu>=1.8.0

# RAG utilities
tiktoken>=0.7.0
bs4>=0.0.2
requests>=2.32.3

# UI
streamlit>=1.38.0

# Environment handling
python-dotenv>=1.0.1

# Data science stack
numpy>=1.26.4
pandas>=2.2.2
